set fps to 90; tweak difficulty speeds
Setting to 90 was essential to get the snake speed feeling good for Ultra difficulty.
Each frame at 90fps is 11.111111 ms, so we have smaller intervals to work with.

I did perf testing up to 3000 snake segments and on average `draw` took roughly 6ms
to complete, well below the 11ms cutoff. There were spikes above 18ms (max time), so
this meant there were some framme drops at 3000 snake segments, but 18ms still achieves
60 fps so I'm happy. Besides, 3000 is an order of magnitude more segments than would be
seen in the worst case scenario for gameplay.

Lastly - speeds are deliberately set a little over the number of frames to make the speed
more consistent. For example, Ultra sets the speed to 25, while 2x(1000/90)=22.2222. And
Hard is set to 35, while 3x(1000/90)=33.3333. The reason for this is that we want a
consistent number of frames between snake moves. The frame rate does not stay exactly at
11.1111ms but will contain some indeterminate variance.